move LLR processing into the separate .go file
